Lift app now updates a Woot model, and can send document to clients that join.
TODO:
* Operation needs to store a "from" address (done in Javascript).
* When joining a conversation, we need to send Lift's model's queue

+++ src/main/scala/woots/package.scala
@@ -1,7 +1,7 @@
 
 package object woots {
 
-  type SiteId = Long
-  type ClockValue = Long
+  type SiteId = Int
+  type ClockValue = Int
   
 }
